+++ code/core/src/main/java/io/zabrek/soulbound/config/migrator/from1to2/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Migrations from 1.x to 2.x.
+ */
+package io.zabrek.soulbound.config.migrator.from1to2;