+++ code/core/src/main/java/io/zabrek/soulbound/config/migrator/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Config patcher migration classes.
+ */
+package io.zabrek.soulbound.config.migrator;