+++ code/lib/src/main/java/io/zabrek/soulbound/lib/config/patcher/migration/Migration.java
@@ -1,0 +1,17 @@
+package io.zabrek.soulbound.lib.config.patcher.migration;
+
+import java.io.IOException;
+
+/**
+ * Handles the migration process for generic changes.
+ */
+@FunctionalInterface
+public interface Migration {
+
+    /**
+     * Migrates the configs.
+     *
+     * @throws IOException if an error occurs
+     */
+    void migrate() throws IOException;
+}