+++ code/lib/src/main/java/io/zabrek/soulbound/lib/config/patcher/migration/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Configuration migration classes.
+ */
+package io.zabrek.soulbound.lib.config.patcher.migration;